fix: renaming reduxStore to store
