+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 
-import store from './redux/reduxStore';
+import store from './redux/store';
 import App from './App';
 
 import './index.css';
